VK: refactoring part 1
