+++ rpcs3/Emu/RSX/VK/VKFragmentProgram.h
@@ -3,7 +3,8 @@
 #include "../Common/GLSLTypes.h"
 #include "Emu/RSX/RSXFragmentProgram.h"
 #include "VulkanAPI.h"
-#include "VKHelpers.h"
+#include "VKProgramPipeline.h"
+#include "helpers/pipeline_binding_table.h"
 
 namespace vk
 {
