+++ rpcs3/Emu/RSX/VK/VKHelpers.h
@@ -21,12 +21,14 @@
 #include "../Common/TextureUtils.h"
 #include "../display.h"
 #include "../rsx_utils.h"
-
-#include "3rdparty/GPUOpen/include/vk_mem_alloc.h"
-
-#ifdef _MSC_VER
-extern "C" void _mm_pause();
-#endif
+#include "helpers/chip_class.h"
+#include "helpers/fence.h"
+#include "helpers/mem_allocator.h"
+#include "helpers/memory_block.h"
+#include "helpers/physical_device.h"
+#include "helpers/pipeline_binding_table.h"
+#include "helpers/shared.h"
+#include "helpers/supported_extensions.h"
 
 #ifdef __APPLE__
 #define VK_DISABLE_COMPONENT_SWIZZLE 1
@@ -50,15 +52,9 @@ namespace rsx
 
 namespace vk
 {
-#define CHECK_RESULT(expr) { VkResult _res = (expr); if (_res != VK_SUCCESS) vk::die_with_error(_res); }
-
 	VKAPI_ATTR void *VKAPI_CALL mem_realloc(void *pUserData, void *pOriginal, usz size, usz alignment, VkSystemAllocationScope allocationScope);
 	VKAPI_ATTR void *VKAPI_CALL mem_alloc(void *pUserData, usz size, usz alignment, VkSystemAllocationScope allocationScope);
 	VKAPI_ATTR void VKAPI_CALL mem_free(void *pUserData, void *pMemory);
-
-	VKAPI_ATTR VkBool32 VKAPI_CALL dbgFunc(VkFlags msgFlags, VkDebugReportObjectTypeEXT objType,
-											u64 srcObject, usz location, s32 msgCode,
-											const char *pLayerPrefix, const char *pMsg, void *pUserData);
 
 	VkBool32 BreakCallback(VkFlags msgFlags, VkDebugReportObjectTypeEXT objType,
 							u64 srcObject, usz location, s32 msgCode,
@@ -73,58 +69,20 @@ namespace vk
 		heap_changed = 3
 	};
 
-	enum class driver_vendor
-	{
-		unknown,
-		AMD,
-		NVIDIA,
-		RADV,
-		INTEL
-	};
-
-	// Chip classes grouped by vendor in order of release
-	enum class chip_class
-	{
-		unknown,
-		AMD_gcn_generic,
-		AMD_polaris,
-		AMD_vega,
-		AMD_navi1x,
-		AMD_navi2x,
-		NV_generic,
-		NV_kepler,
-		NV_maxwell,
-		NV_pascal,
-		NV_volta,
-		NV_turing,
-		NV_ampere
-	};
-
 	enum : u32// special remap_encoding enums
 	{
 		VK_REMAP_IDENTITY = 0xCAFEBABE,             // Special view encoding to return an identity image view
 		VK_REMAP_VIEW_MULTISAMPLED = 0xDEADBEEF     // Special encoding for multisampled images; returns a multisampled image view
 	};
 
-	enum // callback commands
-	{
-		rctrl_queue_submit = 0x80000000
-	};
-
 	class context;
 	class render_device;
 	class swap_chain_image;
-	class physical_device;
 	class command_buffer;
 	class image;
 	struct image_view;
 	struct buffer;
 	class data_heap;
-	class mem_allocator_base;
-	struct memory_type_mapping;
-	struct gpu_formats_support;
-	struct fence;
-	struct pipeline_binding_table;
 	class event;
 
 	const vk::context *get_current_thread_ctx();
@@ -133,8 +91,6 @@ namespace vk
 	const vk::render_device *get_current_renderer();
 	void set_current_renderer(const vk::render_device &device);
 
-	mem_allocator_base *get_current_mem_allocator();
-
 	//Compatibility workarounds
 	bool emulate_primitive_restart(rsx::primitive_type type);
 	bool sanitize_fp_values();
@@ -142,8 +98,6 @@ namespace vk
 	bool emulate_conditional_rendering();
 	VkFlags get_heap_compatible_buffer_types();
 	driver_vendor get_driver_vendor();
-	chip_class get_chip_family(u32 vendor_id, u32 device_id);
-	chip_class get_chip_family();
 
 	VkComponentMapping default_component_map();
 	VkComponentMapping apply_swizzle_remap(const std::array<VkComponentSwizzle, 4>& base_remap, const std::pair<std::array<u8, 4>, std::array<u8, 4>>& remap_vector);
@@ -156,10 +110,6 @@ namespace vk
 	image* get_typeless_helper(VkFormat format, rsx::format_class format_class, u32 requested_width, u32 requested_height);
 	buffer* get_scratch_buffer(u32 min_required_size = 0);
 	data_heap* get_upload_heap();
-
-	memory_type_mapping get_memory_mapping(const physical_device& dev);
-	gpu_formats_support get_optimal_tiling_supported_formats(const physical_device& dev);
-	pipeline_binding_table get_pipeline_binding_table(const physical_device& dev);
 
 	// Sync helpers around vkQueueSubmit
 	void acquire_global_submit_lock();
@@ -175,12 +125,6 @@ namespace vk
 
 	void destroy_global_resources();
 	void reset_global_resources();
-
-	void vmm_notify_memory_allocated(void* handle, u32 memory_type, u64 memory_size);
-	void vmm_notify_memory_freed(void* handle);
-	void vmm_reset();
-	void vmm_check_memory_usage();
-	bool vmm_handle_memory_pressure(rsx::problem_severity severity);
 
 	/**
 	* Allocate enough space in upload_buffer and write all mipmap/layer data into the subbuffer.
@@ -247,636 +191,6 @@ namespace vk
 	// Handle unexpected submit with dangling occlusion query
 	// TODO: Move queries out of the renderer!
 	void do_query_cleanup(vk::command_buffer& cmd);
-
-	void die_with_error(VkResult error_code,
-		const char* file = __builtin_FILE(),
-		const char* func = __builtin_FUNCTION(),
-		u32 line = __builtin_LINE(),
-		u32 col = __builtin_COLUMN());
-
-	struct pipeline_binding_table
-	{
-		u8 vertex_params_bind_slot = 0;
-		u8 vertex_constant_buffers_bind_slot = 1;
-		u8 fragment_constant_buffers_bind_slot = 2;
-		u8 fragment_state_bind_slot = 3;
-		u8 fragment_texture_params_bind_slot = 4;
-		u8 vertex_buffers_first_bind_slot = 5;
-		u8 conditional_render_predicate_slot = 8;
-		u8 rasterizer_env_bind_slot = 9;
-		u8 textures_first_bind_slot = 10;
-		u8 vertex_textures_first_bind_slot = 10;  // Invalid, has to be initialized properly
-		u8 total_descriptor_bindings = vertex_textures_first_bind_slot; // Invalid, has to be initialized properly
-	};
-
-	struct memory_type_mapping
-	{
-		u32 host_visible_coherent;
-		u32 device_local;
-	};
-
-	struct gpu_formats_support
-	{
-		bool d24_unorm_s8;
-		bool d32_sfloat_s8;
-		bool bgra8_linear;
-		bool argb8_linear;
-	};
-
-	struct gpu_shader_types_support
-	{
-		bool allow_float64;
-		bool allow_float16;
-		bool allow_int8;
-	};
-
-	struct chip_family_table
-	{
-		chip_class default_ = chip_class::unknown;
-		std::unordered_map<u32, chip_class> lut;
-
-		void add(u32 first, u32 last, chip_class family)
-		{
-			for (auto i = first; i <= last; ++i)
-			{
-				lut[i] = family;
-			}
-		}
-
-		void add(u32 id, chip_class family)
-		{
-			lut[id] = family;
-		}
-
-		chip_class find(u32 device_id)
-		{
-			if (auto found = lut.find(device_id); found != lut.end())
-			{
-				return found->second;
-			}
-
-			rsx_log.warning("Unknown chip with device ID 0x%x", device_id);
-			return default_;
-		}
-	};
-
-	// Memory Allocator - base class
-
-	class mem_allocator_base
-	{
-	public:
-		using mem_handle_t = void *;
-
-		mem_allocator_base(VkDevice dev, VkPhysicalDevice /*pdev*/) : m_device(dev) {}
-		virtual ~mem_allocator_base() = default;
-
-		virtual void destroy() = 0;
-
-		virtual mem_handle_t alloc(u64 block_sz, u64 alignment, u32 memory_type_index) = 0;
-		virtual void free(mem_handle_t mem_handle) = 0;
-		virtual void *map(mem_handle_t mem_handle, u64 offset, u64 size) = 0;
-		virtual void unmap(mem_handle_t mem_handle) = 0;
-		virtual VkDeviceMemory get_vk_device_memory(mem_handle_t mem_handle) = 0;
-		virtual u64 get_vk_device_memory_offset(mem_handle_t mem_handle) = 0;
-		virtual f32 get_memory_usage() = 0;
-
-	protected:
-		VkDevice m_device;
-	private:
-	};
-
-	// Memory Allocator - Vulkan Memory Allocator
-	// https://github.com/GPUOpen-LibrariesAndSDKs/VulkanMemoryAllocator
-
-	class mem_allocator_vma : public mem_allocator_base
-	{
-	public:
-		mem_allocator_vma(VkDevice dev, VkPhysicalDevice pdev) : mem_allocator_base(dev, pdev)
-		{
-			// Initialize stats pool
-			std::fill(stats.begin(), stats.end(), VmaBudget{});
-
-			VmaAllocatorCreateInfo allocatorInfo = {};
-			allocatorInfo.physicalDevice = pdev;
-			allocatorInfo.device = dev;
-
-			vmaCreateAllocator(&allocatorInfo, &m_allocator);
-		}
-
-		~mem_allocator_vma() override = default;
-
-		void destroy() override
-		{
-			vmaDestroyAllocator(m_allocator);
-		}
-
-		mem_handle_t alloc(u64 block_sz, u64 alignment, u32 memory_type_index) override
-		{
-			VmaAllocation vma_alloc;
-			VkMemoryRequirements mem_req = {};
-			VmaAllocationCreateInfo create_info = {};
-
-			mem_req.memoryTypeBits = 1u << memory_type_index;
-			mem_req.size = block_sz;
-			mem_req.alignment = alignment;
-			create_info.memoryTypeBits = 1u << memory_type_index;
-
-			if (VkResult result = vmaAllocateMemory(m_allocator, &mem_req, &create_info, &vma_alloc, nullptr);
-				result != VK_SUCCESS)
-			{
-				if (result == VK_ERROR_OUT_OF_DEVICE_MEMORY &&
-					vmm_handle_memory_pressure(rsx::problem_severity::fatal))
-				{
-					// If we just ran out of VRAM, attempt to release resources and try again
-					result = vmaAllocateMemory(m_allocator, &mem_req, &create_info, &vma_alloc, nullptr);
-				}
-
-				if (result != VK_SUCCESS)
-				{
-					die_with_error(result);
-				}
-				else
-				{
-					rsx_log.warning("Renderer ran out of video memory but successfully recovered.");
-				}
-			}
-
-			vmm_notify_memory_allocated(vma_alloc, memory_type_index, block_sz);
-			return vma_alloc;
-		}
-
-		void free(mem_handle_t mem_handle) override
-		{
-			vmm_notify_memory_freed(mem_handle);
-			vmaFreeMemory(m_allocator, static_cast<VmaAllocation>(mem_handle));
-		}
-
-		void *map(mem_handle_t mem_handle, u64 offset, u64 /*size*/) override
-		{
-			void *data = nullptr;
-
-			CHECK_RESULT(vmaMapMemory(m_allocator, static_cast<VmaAllocation>(mem_handle), &data));
-
-			// Add offset
-			data = static_cast<u8 *>(data) + offset;
-			return data;
-		}
-
-		void unmap(mem_handle_t mem_handle) override
-		{
-			vmaUnmapMemory(m_allocator, static_cast<VmaAllocation>(mem_handle));
-		}
-
-		VkDeviceMemory get_vk_device_memory(mem_handle_t mem_handle) override
-		{
-			VmaAllocationInfo alloc_info;
-
-			vmaGetAllocationInfo(m_allocator, static_cast<VmaAllocation>(mem_handle), &alloc_info);
-			return alloc_info.deviceMemory;
-		}
-
-		u64 get_vk_device_memory_offset(mem_handle_t mem_handle) override
-		{
-			VmaAllocationInfo alloc_info;
-
-			vmaGetAllocationInfo(m_allocator, static_cast<VmaAllocation>(mem_handle), &alloc_info);
-			return alloc_info.offset;
-		}
-
-		f32 get_memory_usage() override
-		{
-			vmaGetBudget(m_allocator, stats.data());
-
-			float max_usage = 0.f;
-			for (const auto& info : stats)
-			{
-				if (!info.budget)
-				{
-					break;
-				}
-
-				const float this_usage = (info.usage * 100.f) / info.budget;
-				max_usage = std::max(max_usage, this_usage);
-			}
-
-			return max_usage;
-		}
-
-	private:
-		VmaAllocator m_allocator;
-		std::array<VmaBudget, VK_MAX_MEMORY_HEAPS> stats;
-	};
-
-	// Memory Allocator - built-in Vulkan device memory allocate/free
-
-	class mem_allocator_vk : public mem_allocator_base
-	{
-	public:
-		mem_allocator_vk(VkDevice dev, VkPhysicalDevice pdev) : mem_allocator_base(dev, pdev) {}
-		~mem_allocator_vk() override = default;
-
-		void destroy() override {}
-
-		mem_handle_t alloc(u64 block_sz, u64 /*alignment*/, u32 memory_type_index) override
-		{
-			VkDeviceMemory memory;
-			VkMemoryAllocateInfo info = {};
-			info.sType = VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO;
-			info.allocationSize = block_sz;
-			info.memoryTypeIndex = memory_type_index;
-
-			if (VkResult result = vkAllocateMemory(m_device, &info, nullptr, &memory);
-				result != VK_SUCCESS)
-			{
-				if (result == VK_ERROR_OUT_OF_DEVICE_MEMORY &&
-					vmm_handle_memory_pressure(rsx::problem_severity::fatal))
-				{
-					// If we just ran out of VRAM, attempt to release resources and try again
-					result = vkAllocateMemory(m_device, &info, nullptr, &memory);
-				}
-
-				if (result != VK_SUCCESS)
-				{
-					die_with_error(result);
-				}
-				else
-				{
-					rsx_log.warning("Renderer ran out of video memory but successfully recovered.");
-				}
-			}
-
-			vmm_notify_memory_allocated(memory, memory_type_index, block_sz);
-			return memory;
-		}
-
-		void free(mem_handle_t mem_handle) override
-		{
-			vmm_notify_memory_freed(mem_handle);
-			vkFreeMemory(m_device, static_cast<VkDeviceMemory>(mem_handle), nullptr);
-		}
-
-		void *map(mem_handle_t mem_handle, u64 offset, u64 size) override
-		{
-			void *data = nullptr;
-			CHECK_RESULT(vkMapMemory(m_device, static_cast<VkDeviceMemory>(mem_handle), offset, std::max<u64>(size, 1u), 0, &data));
-			return data;
-		}
-
-		void unmap(mem_handle_t mem_handle) override
-		{
-			vkUnmapMemory(m_device, static_cast<VkDeviceMemory>(mem_handle));
-		}
-
-		VkDeviceMemory get_vk_device_memory(mem_handle_t mem_handle) override
-		{
-			return static_cast<VkDeviceMemory>(mem_handle);
-		}
-
-		u64 get_vk_device_memory_offset(mem_handle_t /*mem_handle*/) override
-		{
-			return 0;
-		}
-
-		f32 get_memory_usage() override
-		{
-			return 0.f;
-		}
-
-	private:
-	};
-
-	struct memory_block
-	{
-		memory_block(VkDevice dev, u64 block_sz, u64 alignment, u32 memory_type_index) : m_device(dev)
-		{
-			m_mem_allocator = get_current_mem_allocator();
-			m_mem_handle = m_mem_allocator->alloc(block_sz, alignment, memory_type_index);
-		}
-
-		~memory_block()
-		{
-			m_mem_allocator->free(m_mem_handle);
-		}
-
-		VkDeviceMemory get_vk_device_memory()
-		{
-			return m_mem_allocator->get_vk_device_memory(m_mem_handle);
-		}
-
-		u64 get_vk_device_memory_offset()
-		{
-			return m_mem_allocator->get_vk_device_memory_offset(m_mem_handle);
-		}
-
-		void *map(u64 offset, u64 size)
-		{
-			return m_mem_allocator->map(m_mem_handle, offset, size);
-		}
-
-		void unmap()
-		{
-			m_mem_allocator->unmap(m_mem_handle);
-		}
-
-		memory_block(const memory_block&) = delete;
-		memory_block(memory_block&&) = delete;
-
-	private:
-		VkDevice m_device;
-		vk::mem_allocator_base* m_mem_allocator;
-		mem_allocator_base::mem_handle_t m_mem_handle;
-	};
-
-	class supported_extensions
-	{
-	private:
-		std::vector<VkExtensionProperties> m_vk_exts;
-
-	public:
-		enum enumeration_class
-		{
-			instance = 0,
-			device = 1
-		};
-
-		supported_extensions(enumeration_class _class, const char* layer_name = nullptr, VkPhysicalDevice pdev = VK_NULL_HANDLE)
-		{
-			u32 count;
-			if (_class == enumeration_class::instance)
-			{
-				if (vkEnumerateInstanceExtensionProperties(layer_name, &count, nullptr) != VK_SUCCESS)
-					return;
-			}
-			else
-			{
-				ensure(pdev);
-				if (vkEnumerateDeviceExtensionProperties(pdev, layer_name, &count, nullptr) != VK_SUCCESS)
-					return;
-			}
-
-			m_vk_exts.resize(count);
-			if (_class == enumeration_class::instance)
-			{
-				vkEnumerateInstanceExtensionProperties(layer_name, &count, m_vk_exts.data());
-			}
-			else
-			{
-				vkEnumerateDeviceExtensionProperties(pdev, layer_name, &count, m_vk_exts.data());
-			}
-		}
-
-		bool is_supported(const char *ext)
-		{
-			return std::any_of(m_vk_exts.cbegin(), m_vk_exts.cend(),
-				[&](const VkExtensionProperties& p) { return std::strcmp(p.extensionName, ext) == 0; });
-		}
-	};
-
-	class physical_device
-	{
-		VkInstance parent = VK_NULL_HANDLE;
-		VkPhysicalDevice dev = VK_NULL_HANDLE;
-		VkPhysicalDeviceProperties props;
-		VkPhysicalDeviceFeatures features;
-		VkPhysicalDeviceMemoryProperties memory_properties;
-		std::vector<VkQueueFamilyProperties> queue_props;
-
-		std::unordered_map<VkFormat, VkFormatProperties> format_properties;
-		gpu_shader_types_support shader_types_support{};
-		VkPhysicalDeviceDriverPropertiesKHR driver_properties{};
-
-		bool stencil_export_support = false;
-		bool conditional_render_support = false;
-		bool unrestricted_depth_range_support = false;
-
-		friend class render_device;
-private:
-		void get_physical_device_features(bool allow_extensions)
-		{
-			if (!allow_extensions)
-			{
-				vkGetPhysicalDeviceFeatures(dev, &features);
-				return;
-			}
-
-			supported_extensions instance_extensions(supported_extensions::instance);
-			supported_extensions device_extensions(supported_extensions::device, nullptr, dev);
-
-			if (!instance_extensions.is_supported(VK_KHR_GET_PHYSICAL_DEVICE_PROPERTIES_2_EXTENSION_NAME))
-			{
-				vkGetPhysicalDeviceFeatures(dev, &features);
-			}
-			else
-			{
-				VkPhysicalDeviceFeatures2KHR features2;
-				features2.sType = VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FEATURES_2;
-				features2.pNext = nullptr;
-
-				VkPhysicalDeviceFloat16Int8FeaturesKHR shader_support_info{};
-
-				if (device_extensions.is_supported(VK_KHR_SHADER_FLOAT16_INT8_EXTENSION_NAME))
-				{
-					shader_support_info.sType = VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FLOAT16_INT8_FEATURES_KHR;
-					features2.pNext = &shader_support_info;
-				}
-
-				if (device_extensions.is_supported(VK_KHR_DRIVER_PROPERTIES_EXTENSION_NAME))
-				{
-					driver_properties.sType = VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DRIVER_PROPERTIES_KHR;
-					driver_properties.pNext = features2.pNext;
-					features2.pNext = &driver_properties;
-				}
-
-				auto getPhysicalDeviceFeatures2KHR = reinterpret_cast<PFN_vkGetPhysicalDeviceFeatures2KHR>(vkGetInstanceProcAddr(parent, "vkGetPhysicalDeviceFeatures2KHR"));
-				ensure(getPhysicalDeviceFeatures2KHR); // "vkGetInstanceProcAddress failed to find entry point!"
-				getPhysicalDeviceFeatures2KHR(dev, &features2);
-
-				shader_types_support.allow_float64 = !!features2.features.shaderFloat64;
-				shader_types_support.allow_float16 = !!shader_support_info.shaderFloat16;
-				shader_types_support.allow_int8 = !!shader_support_info.shaderInt8;
-				features = features2.features;
-			}
-
-			stencil_export_support = device_extensions.is_supported(VK_EXT_SHADER_STENCIL_EXPORT_EXTENSION_NAME);
-			conditional_render_support = device_extensions.is_supported(VK_EXT_CONDITIONAL_RENDERING_EXTENSION_NAME);
-			unrestricted_depth_range_support = device_extensions.is_supported(VK_EXT_DEPTH_RANGE_UNRESTRICTED_EXTENSION_NAME);
-		}
-
-	public:
-
-		physical_device() = default;
-		~physical_device() = default;
-
-		void create(VkInstance context, VkPhysicalDevice pdev, bool allow_extensions)
-		{
-			dev = pdev;
-			parent = context;
-			vkGetPhysicalDeviceProperties(pdev, &props);
-			vkGetPhysicalDeviceMemoryProperties(pdev, &memory_properties);
-			get_physical_device_features(allow_extensions);
-
-			rsx_log.always("Found vulkan-compatible GPU: '%s' running on driver %s", get_name(), get_driver_version());
-
-			if (get_driver_vendor() == driver_vendor::RADV &&
-				get_name().find("LLVM 8.0.0") != umax)
-			{
-				// Serious driver bug causing black screens
-				// See https://bugs.freedesktop.org/show_bug.cgi?id=110970
-				rsx_log.fatal("RADV drivers have a major driver bug with LLVM 8.0.0 resulting in no visual output. Upgrade to LLVM version 8.0.1 or greater to avoid this issue.");
-			}
-			else if (get_driver_vendor() == driver_vendor::NVIDIA)
-			{
-#ifdef _WIN32
-				// SPIRV bugs were fixed in 452.28 for windows
-				const u32 threshold_version = (452u >> 22) | (28 >> 14);
-#else
-				// SPIRV bugs were fixed in 450.56 for linux/BSD
-				const u32 threshold_version = (450u >> 22) | (56 >> 14);
-#endif
-				const auto current_version = props.driverVersion & ~0x3fffu; // Clear patch and revision fields
-				if (current_version < threshold_version)
-				{
-					rsx_log.error("Your current NVIDIA graphics driver version %s has known issues and is unsupported. Update to the latest NVIDIA driver.",
-						get_driver_version());
-				}
-			}
-
-			if (get_chip_class() == chip_class::AMD_vega)
-			{
-				// Disable fp16 if driver uses LLVM emitter. It does fine with AMD proprietary drivers though.
-				shader_types_support.allow_float16 = (driver_properties.driverID == VK_DRIVER_ID_AMD_PROPRIETARY_KHR);
-			}
-		}
-
-		std::string get_name() const
-		{
-			return props.deviceName;
-		}
-
-		driver_vendor get_driver_vendor() const
-		{
-			if (!driver_properties.driverID)
-			{
-				const auto gpu_name = get_name();
-				if (gpu_name.find("Radeon") != umax)
-				{
-					return driver_vendor::AMD;
-				}
-
-				if (gpu_name.find("NVIDIA") != umax || gpu_name.find("GeForce") != umax || gpu_name.find("Quadro") != umax)
-				{
-					return driver_vendor::NVIDIA;
-				}
-
-				if (gpu_name.find("RADV") != umax)
-				{
-					return driver_vendor::RADV;
-				}
-
-				if (gpu_name.find("Intel") != umax)
-				{
-					return driver_vendor::INTEL;
-				}
-
-				return driver_vendor::unknown;
-			}
-			else
-			{
-				switch (driver_properties.driverID)
-				{
-				case VK_DRIVER_ID_AMD_PROPRIETARY_KHR:
-				case VK_DRIVER_ID_AMD_OPEN_SOURCE_KHR:
-					return driver_vendor::AMD;
-				case VK_DRIVER_ID_MESA_RADV_KHR:
-					return driver_vendor::RADV;
-				case VK_DRIVER_ID_NVIDIA_PROPRIETARY_KHR:
-					return driver_vendor::NVIDIA;
-				case VK_DRIVER_ID_INTEL_PROPRIETARY_WINDOWS_KHR:
-				case VK_DRIVER_ID_INTEL_OPEN_SOURCE_MESA_KHR:
-					return driver_vendor::INTEL;
-				default:
-					// Mobile
-					return driver_vendor::unknown;
-				}
-			}
-		}
-
-		std::string get_driver_version() const
-		{
-			switch (get_driver_vendor())
-			{
-			case driver_vendor::NVIDIA:
-			{
-				// 10 + 8 + 8 + 6
-				const auto major_version = props.driverVersion >> 22;
-				const auto minor_version = (props.driverVersion >> 14) & 0xff;
-				const auto patch = (props.driverVersion >> 6) & 0xff;
-				const auto revision = (props.driverVersion & 0x3f);
-
-				return fmt::format("%u.%u.%u.%u", major_version, minor_version, patch, revision);
-			}
-			default:
-			{
-				// 10 + 10 + 12 (standard vulkan encoding created with VK_MAKE_VERSION)
-				return fmt::format("%u.%u.%u",
-					(props.driverVersion >> 22),
-					(props.driverVersion >> 12) & 0x3ff,
-					(props.driverVersion) & 0x3ff);
-			}
-			}
-		}
-
-		chip_class get_chip_class() const
-		{
-			return get_chip_family(props.vendorID, props.deviceID);
-		}
-
-		u32 get_queue_count() const
-		{
-			if (!queue_props.empty())
-				return ::size32(queue_props);
-
-			u32 count = 0;
-			vkGetPhysicalDeviceQueueFamilyProperties(dev, &count, nullptr);
-
-			return count;
-		}
-
-		VkQueueFamilyProperties get_queue_properties(u32 queue)
-		{
-			if (queue_props.empty())
-			{
-				u32 count = 0;
-				vkGetPhysicalDeviceQueueFamilyProperties(dev, &count, nullptr);
-
-				queue_props.resize(count);
-				vkGetPhysicalDeviceQueueFamilyProperties(dev, &count, queue_props.data());
-			}
-
-			if (queue >= queue_props.size()) fmt::throw_exception("Bad queue index passed to get_queue_properties (%u)", queue);
-			return queue_props[queue];
-		}
-
-		VkPhysicalDeviceMemoryProperties get_memory_properties() const
-		{
-			return memory_properties;
-		}
-
-		VkPhysicalDeviceLimits get_limits() const
-		{
-			return props.limits;
-		}
-
-		operator VkPhysicalDevice() const
-		{
-			return dev;
-		}
-
-		operator VkInstance() const
-		{
-			return parent;
-		}
-	};
 
 	class render_device
 	{
@@ -1194,58 +508,6 @@ private:
 		operator VkCommandPool()
 		{
 			return pool;
-		}
-	};
-
-	struct fence
-	{
-		atomic_t<bool> flushed = false;
-		VkFence handle         = VK_NULL_HANDLE;
-		VkDevice owner         = VK_NULL_HANDLE;
-
-		fence(VkDevice dev)
-		{
-			owner = dev;
-			VkFenceCreateInfo info = {};
-			info.sType = VK_STRUCTURE_TYPE_FENCE_CREATE_INFO;
-			CHECK_RESULT(vkCreateFence(dev, &info, nullptr, &handle));
-		}
-
-		~fence()
-		{
-			if (handle)
-			{
-				vkDestroyFence(owner, handle, nullptr);
-				handle = VK_NULL_HANDLE;
-			}
-		}
-
-		void reset()
-		{
-			vkResetFences(owner, 1, &handle);
-			flushed.release(false);
-		}
-
-		void signal_flushed()
-		{
-			flushed.release(true);
-		}
-
-		void wait_flush()
-		{
-			while (!flushed)
-			{
-#ifdef _MSC_VER
-				_mm_pause();
-#else
-				__builtin_ia32_pause();
-#endif
-			}
-		}
-
-		operator bool() const
-		{
-			return (handle != VK_NULL_HANDLE);
 		}
 	};
 
@@ -1988,64 +1250,6 @@ private:
 				return (*m_value == 0xDEADBEEF)? VK_EVENT_SET : VK_EVENT_RESET;
 			}
 		}
-	};
-
-	struct sampler
-	{
-		VkSampler value;
-		VkSamplerCreateInfo info = {};
-
-		sampler(VkDevice dev, VkSamplerAddressMode clamp_u, VkSamplerAddressMode clamp_v, VkSamplerAddressMode clamp_w,
-			VkBool32 unnormalized_coordinates, float mipLodBias, float max_anisotropy, float min_lod, float max_lod,
-			VkFilter min_filter, VkFilter mag_filter, VkSamplerMipmapMode mipmap_mode, VkBorderColor border_color,
-			VkBool32 depth_compare = false, VkCompareOp depth_compare_mode = VK_COMPARE_OP_NEVER)
-			: m_device(dev)
-		{
-			info.sType = VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO;
-			info.addressModeU = clamp_u;
-			info.addressModeV = clamp_v;
-			info.addressModeW = clamp_w;
-			info.anisotropyEnable = VK_TRUE;
-			info.compareEnable = depth_compare;
-			info.unnormalizedCoordinates = unnormalized_coordinates;
-			info.mipLodBias = mipLodBias;
-			info.maxAnisotropy = max_anisotropy;
-			info.maxLod = max_lod;
-			info.minLod = min_lod;
-			info.magFilter = mag_filter;
-			info.minFilter = min_filter;
-			info.mipmapMode = mipmap_mode;
-			info.compareOp = depth_compare_mode;
-			info.borderColor = border_color;
-
-			CHECK_RESULT(vkCreateSampler(m_device, &info, nullptr, &value));
-		}
-
-		~sampler()
-		{
-			vkDestroySampler(m_device, value, nullptr);
-		}
-
-		bool matches(VkSamplerAddressMode clamp_u, VkSamplerAddressMode clamp_v, VkSamplerAddressMode clamp_w,
-			VkBool32 unnormalized_coordinates, float mipLodBias, float max_anisotropy, float min_lod, float max_lod,
-			VkFilter min_filter, VkFilter mag_filter, VkSamplerMipmapMode mipmap_mode, VkBorderColor border_color,
-			VkBool32 depth_compare = false, VkCompareOp depth_compare_mode = VK_COMPARE_OP_NEVER)
-		{
-			if (info.magFilter != mag_filter || info.minFilter != min_filter || info.mipmapMode != mipmap_mode ||
-				info.addressModeU != clamp_u || info.addressModeV != clamp_v || info.addressModeW != clamp_w ||
-				info.compareEnable != depth_compare || info.unnormalizedCoordinates != unnormalized_coordinates ||
-				!rsx::fcmp(info.maxLod, max_lod) || !rsx::fcmp(info.mipLodBias, mipLodBias) || !rsx::fcmp(info.minLod, min_lod) ||
-				!rsx::fcmp(info.maxAnisotropy, max_anisotropy) ||
-				info.compareOp != depth_compare_mode || info.borderColor != border_color)
-				return false;
-
-			return true;
-		}
-
-		sampler(const sampler&) = delete;
-		sampler(sampler&&) = delete;
-	private:
-		VkDevice m_device;
 	};
 
 	struct framebuffer
@@ -3286,399 +2490,6 @@ public:
 			CHECK_RESULT(vkResetDescriptorPool(*m_owner, m_current_pool_handle, flags));
 		}
 	};
-
-	class query_pool : public rsx::ref_counted
-	{
-		VkQueryPool m_query_pool;
-		VkDevice m_device;
-
-	public:
-		query_pool(VkDevice dev, VkQueryType type, u32 size)
-			: m_query_pool(VK_NULL_HANDLE), m_device(dev)
-		{
-			VkQueryPoolCreateInfo info{};
-			info.sType = VK_STRUCTURE_TYPE_QUERY_POOL_CREATE_INFO;
-			info.queryType = type;
-			info.queryCount = size;
-			vkCreateQueryPool(dev, &info, nullptr, &m_query_pool);
-
-			// Take 'size' references on this object
-			ref_count.release(static_cast<s32>(size));
-		}
-
-		~query_pool()
-		{
-			vkDestroyQueryPool(m_device, m_query_pool, nullptr);
-		}
-
-		operator VkQueryPool()
-		{
-			return m_query_pool;
-		}
-	};
-
-	class graphics_pipeline_state
-	{
-	public:
-		VkPipelineInputAssemblyStateCreateInfo ia;
-		VkPipelineDepthStencilStateCreateInfo ds;
-		VkPipelineColorBlendAttachmentState att_state[4];
-		VkPipelineColorBlendStateCreateInfo cs;
-		VkPipelineRasterizationStateCreateInfo rs;
-		VkPipelineMultisampleStateCreateInfo ms;
-
-		struct extra_parameters
-		{
-			VkSampleMask msaa_sample_mask;
-		}
-		temp_storage;
-
-		graphics_pipeline_state()
-		{
-			// NOTE: Vk** structs have padding bytes
-			memset(this, 0, sizeof(graphics_pipeline_state));
-
-			ia.sType = VK_STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO;
-			cs.sType = VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO;
-			ds.sType = VK_STRUCTURE_TYPE_PIPELINE_DEPTH_STENCIL_STATE_CREATE_INFO;
-
-			rs.sType = VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_CREATE_INFO;
-			rs.polygonMode = VK_POLYGON_MODE_FILL;
-			rs.cullMode = VK_CULL_MODE_NONE;
-			rs.frontFace = VK_FRONT_FACE_COUNTER_CLOCKWISE;
-			rs.lineWidth = 1.f;
-
-			ms.sType = VK_STRUCTURE_TYPE_PIPELINE_MULTISAMPLE_STATE_CREATE_INFO;
-			ms.rasterizationSamples = VK_SAMPLE_COUNT_1_BIT;
-			temp_storage.msaa_sample_mask = 0xFFFFFFFF;
-		}
-
-		graphics_pipeline_state(const graphics_pipeline_state& other)
-		{
-			// NOTE: Vk** structs have padding bytes
-			memcpy(this, &other, sizeof(graphics_pipeline_state));
-
-			if (other.cs.pAttachments == other.att_state)
-			{
-				// Rebase pointer
-				cs.pAttachments = att_state;
-			}
-		}
-
-		~graphics_pipeline_state() = default;
-
-		graphics_pipeline_state& operator = (const graphics_pipeline_state& other)
-		{
-			if (this != &other)
-			{
-				// NOTE: Vk** structs have padding bytes
-				memcpy(this, &other, sizeof(graphics_pipeline_state));
-
-				if (other.cs.pAttachments == other.att_state)
-				{
-					// Rebase pointer
-					cs.pAttachments = att_state;
-				}
-			}
-
-			return *this;
-		}
-
-		void set_primitive_type(VkPrimitiveTopology type)
-		{
-			ia.topology = type;
-		}
-
-		void enable_primitive_restart(bool enable = true)
-		{
-			ia.primitiveRestartEnable = enable? VK_TRUE : VK_FALSE;
-		}
-
-		void set_color_mask(int index, bool r, bool g, bool b, bool a)
-		{
-			VkColorComponentFlags mask = 0;
-			if (a) mask |= VK_COLOR_COMPONENT_A_BIT;
-			if (b) mask |= VK_COLOR_COMPONENT_B_BIT;
-			if (g) mask |= VK_COLOR_COMPONENT_G_BIT;
-			if (r) mask |= VK_COLOR_COMPONENT_R_BIT;
-
-			att_state[index].colorWriteMask = mask;
-		}
-
-		void set_depth_mask(bool enable)
-		{
-			ds.depthWriteEnable = enable ? VK_TRUE : VK_FALSE;
-		}
-
-		void set_stencil_mask(u32 mask)
-		{
-			ds.front.writeMask = mask;
-			ds.back.writeMask = mask;
-		}
-
-		void set_stencil_mask_separate(int face, u32 mask)
-		{
-			if (!face)
-				ds.front.writeMask = mask;
-			else
-				ds.back.writeMask = mask;
-		}
-
-		void enable_depth_test(VkCompareOp op)
-		{
-			ds.depthTestEnable = VK_TRUE;
-			ds.depthCompareOp = op;
-		}
-
-		void enable_depth_clamp(bool enable = true)
-		{
-			rs.depthClampEnable = enable ? VK_TRUE : VK_FALSE;
-		}
-
-		void enable_depth_bias(bool enable = true)
-		{
-			rs.depthBiasEnable = enable ? VK_TRUE : VK_FALSE;
-		}
-
-		void enable_depth_bounds_test(bool enable = true)
-		{
-			ds.depthBoundsTestEnable = enable? VK_TRUE : VK_FALSE;
-		}
-
-		void enable_blend(int mrt_index, VkBlendFactor src_factor_rgb, VkBlendFactor src_factor_a,
-			VkBlendFactor dst_factor_rgb, VkBlendFactor dst_factor_a,
-			VkBlendOp equation_rgb, VkBlendOp equation_a)
-		{
-			att_state[mrt_index].srcColorBlendFactor = src_factor_rgb;
-			att_state[mrt_index].srcAlphaBlendFactor = src_factor_a;
-			att_state[mrt_index].dstColorBlendFactor = dst_factor_rgb;
-			att_state[mrt_index].dstAlphaBlendFactor = dst_factor_a;
-			att_state[mrt_index].colorBlendOp = equation_rgb;
-			att_state[mrt_index].alphaBlendOp = equation_a;
-			att_state[mrt_index].blendEnable = VK_TRUE;
-		}
-
-		void enable_stencil_test(VkStencilOp fail, VkStencilOp zfail, VkStencilOp pass,
-			VkCompareOp func, u32 func_mask, u32 ref)
-		{
-			ds.front.failOp = fail;
-			ds.front.passOp = pass;
-			ds.front.depthFailOp = zfail;
-			ds.front.compareOp = func;
-			ds.front.compareMask = func_mask;
-			ds.front.reference = ref;
-			ds.back = ds.front;
-
-			ds.stencilTestEnable = VK_TRUE;
-		}
-
-		void enable_stencil_test_separate(int face, VkStencilOp fail, VkStencilOp zfail, VkStencilOp pass,
-			VkCompareOp func, u32 func_mask, u32 ref)
-		{
-			auto& face_props = (face ? ds.back : ds.front);
-			face_props.failOp = fail;
-			face_props.passOp = pass;
-			face_props.depthFailOp = zfail;
-			face_props.compareOp = func;
-			face_props.compareMask = func_mask;
-			face_props.reference = ref;
-
-			ds.stencilTestEnable = VK_TRUE;
-		}
-
-		void enable_logic_op(VkLogicOp op)
-		{
-			cs.logicOpEnable = VK_TRUE;
-			cs.logicOp = op;
-		}
-
-		void enable_cull_face(VkCullModeFlags cull_mode)
-		{
-			rs.cullMode = cull_mode;
-		}
-
-		void set_front_face(VkFrontFace face)
-		{
-			rs.frontFace = face;
-		}
-
-		void set_attachment_count(u32 count)
-		{
-			cs.attachmentCount = count;
-			cs.pAttachments = att_state;
-		}
-
-		void set_multisample_state(u8 sample_count, u32 sample_mask, bool msaa_enabled, bool alpha_to_coverage, bool alpha_to_one)
-		{
-			temp_storage.msaa_sample_mask = sample_mask;
-
-			ms.rasterizationSamples = static_cast<VkSampleCountFlagBits>(sample_count);
-			ms.alphaToCoverageEnable = alpha_to_coverage;
-			ms.alphaToOneEnable = alpha_to_one;
-
-			if (!msaa_enabled)
-			{
-				// This register is likely glMinSampleShading but in reverse; probably sets max sample shading rate of 1
-				// I (kd-11) suspect its what the control panel setting affects when MSAA is set to disabled
-			}
-		}
-
-		void set_multisample_shading_rate(float shading_rate)
-		{
-			ms.sampleShadingEnable = VK_TRUE;
-			ms.minSampleShading = shading_rate;
-		}
-	};
-
-	namespace glsl
-	{
-		enum program_input_type : u32
-		{
-			input_type_uniform_buffer = 0,
-			input_type_texel_buffer = 1,
-			input_type_texture = 2,
-			input_type_storage_buffer = 3,
-
-			input_type_max_enum = 4
-		};
-
-		struct bound_sampler
-		{
-			VkFormat format;
-			VkImage image;
-			VkComponentMapping mapping;
-		};
-
-		struct bound_buffer
-		{
-			VkFormat format = VK_FORMAT_UNDEFINED;
-			VkBuffer buffer = nullptr;
-			u64 offset = 0;
-			u64 size = 0;
-		};
-
-		struct program_input
-		{
-			::glsl::program_domain domain;
-			program_input_type type;
-
-			bound_buffer as_buffer;
-			bound_sampler as_sampler;
-
-			u32 location;
-			std::string name;
-		};
-
-		class shader
-		{
-			::glsl::program_domain type = ::glsl::program_domain::glsl_vertex_program;
-			VkShaderModule m_handle = VK_NULL_HANDLE;
-			std::string m_source;
-			std::vector<u32> m_compiled;
-
-		public:
-			shader() = default;
-			~shader() = default;
-
-			void create(::glsl::program_domain domain, const std::string& source)
-			{
-				type = domain;
-				m_source = source;
-			}
-
-			VkShaderModule compile()
-			{
-				ensure(m_handle == VK_NULL_HANDLE);
-
-				if (!vk::compile_glsl_to_spv(m_source, type, m_compiled))
-				{
-					std::string shader_type = type == ::glsl::program_domain::glsl_vertex_program ? "vertex" :
-						type == ::glsl::program_domain::glsl_fragment_program ? "fragment" : "compute";
-
-					rsx_log.notice("%s", m_source);
-					fmt::throw_exception("Failed to compile %s shader", shader_type);
-				}
-
-				VkShaderModuleCreateInfo vs_info;
-				vs_info.codeSize = m_compiled.size() * sizeof(u32);
-				vs_info.pNext = nullptr;
-				vs_info.sType = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
-				vs_info.pCode = m_compiled.data();
-				vs_info.flags = 0;
-
-				VkDevice dev = *vk::get_current_renderer();
-				vkCreateShaderModule(dev, &vs_info, nullptr, &m_handle);
-
-				return m_handle;
-			}
-
-			void destroy()
-			{
-				m_source.clear();
-				m_compiled.clear();
-
-				if (m_handle)
-				{
-					VkDevice dev = *vk::get_current_renderer();
-					vkDestroyShaderModule(dev, m_handle, nullptr);
-					m_handle = nullptr;
-				}
-			}
-
-			const std::string& get_source() const
-			{
-				return m_source;
-			}
-
-			const std::vector<u32> get_compiled() const
-			{
-				return m_compiled;
-			}
-
-			VkShaderModule get_handle() const
-			{
-				return m_handle;
-			}
-		};
-
-		class program
-		{
-			std::array<std::vector<program_input>, input_type_max_enum> uniforms;
-			VkDevice m_device;
-
-			std::array<u32, 16> fs_texture_bindings;
-			std::array<u32, 16> fs_texture_mirror_bindings;
-			std::array<u32, 4>  vs_texture_bindings;
-			bool linked;
-
-			void create_impl();
-
-		public:
-			VkPipeline pipeline;
-			VkPipelineLayout pipeline_layout;
-			u64 attribute_location_mask;
-			u64 vertex_attributes_mask;
-
-			program(VkDevice dev, VkPipeline p, VkPipelineLayout layout, const std::vector<program_input> &vertex_input, const std::vector<program_input>& fragment_inputs);
-			program(VkDevice dev, VkPipeline p, VkPipelineLayout layout);
-			program(const program&) = delete;
-			program(program&& other) = delete;
-			~program();
-
-			program& load_uniforms(const std::vector<program_input>& inputs);
-			program& link();
-
-			bool has_uniform(program_input_type type, const std::string &uniform_name);
-			void bind_uniform(const VkDescriptorImageInfo &image_descriptor, const std::string &uniform_name, VkDescriptorType type, VkDescriptorSet &descriptor_set);
-			void bind_uniform(const VkDescriptorImageInfo &image_descriptor, int texture_unit, ::glsl::program_domain domain, VkDescriptorSet &descriptor_set, bool is_stencil_mirror = false);
-			void bind_uniform(const VkDescriptorBufferInfo &buffer_descriptor, u32 binding_point, VkDescriptorSet &descriptor_set);
-			void bind_uniform(const VkBufferView &buffer_view, u32 binding_point, VkDescriptorSet &descriptor_set);
-			void bind_uniform(const VkBufferView &buffer_view, program_input_type type, const std::string &binding_name, VkDescriptorSet &descriptor_set);
-
-			void bind_buffer(const VkDescriptorBufferInfo &buffer_descriptor, u32 binding_point, VkDescriptorType type, VkDescriptorSet &descriptor_set);
-			void bind_descriptor_set(const VkCommandBuffer cmd, VkDescriptorSet descriptor_set);
-		};
-	}
 
 	class data_heap : public ::data_heap
 	{
