+++ rpcs3/Emu/RSX/VK/VKPipelineCompiler.h
@@ -1,11 +1,14 @@
 #pragma once
-#include "VKHelpers.h"
 #include "../rsx_utils.h"
 #include "Utilities/hash.h"
 #include "Utilities/lockless.h"
+#include "VKProgramPipeline.h"
+#include "helpers/graphics_pipeline_state.h"
 
 namespace vk
 {
+	class render_device;
+
 	struct pipeline_props
 	{
 		graphics_pipeline_state state;
